fix forgotten joinsound path renames
